.claude/lab/scripts/requirements.txt
python-docx>=1.1.0
pypdf>=4.0.0
Pillow>=10.0.0
